Move examples to integration tests, using expect instead of console.log
