--- examples/duplicate-message-handling/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "duplicate-message-handling",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "run": "npx ts-node index.ts"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.5"
-  }
-}
