--- examples/duplicate-message-handling/run-example.sh
@@ -1,1 +1,0 @@
-npx ts-node index.ts
